requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-graphql-jwt==0.3.4
django-widget-tweaks==1.4.12
docopt==0.6.2
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
graphqlclient==0.2.4
idna==3.4
Pillow==9.2.0
pipreqs==0.4.11
promise==2.3
PyJWT==2.6.0
python-slugify==6.1.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
urllib3==1.26.12
yarg==0.1.9
